+++ CHANGELOG.md
@@ -16,6 +16,8 @@ v2.11 (Anoia) - (in development)
   - Edit > Mesh > Flip triangles: to flip the triangles (vertices) in case they are defined in the wrong order
 
 - Improvements
+  - Align (Point-pair based registration) tool
+    - can now be used with several entities (both several aligned and several reference entities)
   - Clipping box tool:
     - option to select the extracted contour type (LOWER, UPPER or FULL)
 	  - The 'up' direction is always Z for slices normal to X or Y (the local X, Y or Z directions of the active cross-section)